requirements.txt
streamlit
tensorflow
opencv-python-headless
numpy
keras
gdown
Pillow
